Move demangle into utils
